Updated pagination container.
